# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TruthLens AI Content Detection Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and invalidate token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SCANS
...
  /document/scan:
    post:
      operationId: scanDocument
      tags: [documents]
      summary: Upload document for OCR + AI detection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadBody"
      responses:
        "202":
          description: Document scan job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJobCreated"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /document/{id}:
    get:
      operationId: getDocumentScan
      tags: [documents]
      summary: Get document scan status and results
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentScanResult"
        "404":
          description: Document scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocumentScan
      tags: [documents]
      summary: Delete a document scan record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document scan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /documents:
    get:
      operationId: listDocumentScans
      tags: [documents]
      summary: List user's document scans (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: verdict
          in: query
          schema:
            type: string
            enum: [AI, HUMAN, UNCERTAIN]
      responses:
        "200":
          description: Paginated document scan list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentScanListResponse"

  # STATS / DASHBOARD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        createdAt:
          type: string
          format: date-time
        scanCount:
          type: integer
      required:
        - id
        - email
        - name
        - tier
        - createdAt
        - scanCount
...
    DocumentUploadBody:
      type: object
      properties:
        file:
          type: string
          format: binary
        generateSummary:
          type: boolean
          default: true
        extractKeywords:
          type: boolean
          default: true
      required:
        - file
    ScanJobCreated:
      type: object
      properties:
        scanId:
          type: string
        status:
          type: string
          enum: [pending, processing]
        message:
          type: string
      required:
        - scanId
        - status
        - message
...
    DocumentSection:
      type: object
      properties:
        sectionIndex:
          type: integer
        text:
          type: string
        aiProbability:
          type: number
        highlightColor:
          type: string
          enum: [red, yellow, green]
      required:
        - sectionIndex
        - text
        - aiProbability
        - highlightColor
    DocumentScanResult:
      type: object
      properties:
        docScanId:
          type: string
        status:
          type: string
          enum: [pending, processing, ocr_complete, complete, failed]
        fileName:
          type: string
        confidence:
          type: number
        verdict:
          type: string
          enum: [AI, HUMAN, UNCERTAIN]
        verdictLabel:
          type: string
        ocrText:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSection"
        summary:
          type: string
        keywords:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - docScanId
        - status
        - createdAt
    DocumentScanListItem:
      type: object
      properties:
        docScanId:
          type: string
        fileName:
          type: string
        confidence:
          type: number
        verdict:
          type: string
          enum: [AI, HUMAN, UNCERTAIN]
        verdictLabel:
          type: string
        status:
          type: string
          enum: [pending, processing, ocr_complete, complete, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - docScanId
        - status
        - createdAt
    DocumentScanListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentScanListItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - documents
        - total
        - page
        - limit
        - totalPages
...